requirements.txt
anki==24.4
Pillow==9.3.0
Pillow==10.3.0
